Schedule fortnightly build
Auto-merging dependabot PRs doesn't trigger a `main` build, meaning snapshots are not updated.  The known workaround for this is to use a PAT when doing the merge, but PATs introduce security concerns.

As an alternative to using PATs... schedule a fortnightly build of `main` to ensure snapshots, with updated dependencies, are building built.